pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
requests = "^2.26.0"
gunicorn = "^20.1.0"
pymongo = "^4.2.0"
pastel = "^0.2.1"
Flask-Login = "^0.6.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
mongomock = "^4.1.2"
pytest-watch = "^4.2.0"
selenium = "^4.1.0"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
